requirements.txt
# Core ML and Data Science
numpy>=1.24.0
scipy>=1.10.0
pandas>=1.5.0
scikit-learn>=1.2.0

# Configuration and Data Formats
PyYAML>=6.0.0

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0
tensorflow>=2.13.0

# Computer Vision
opencv-python>=4.8.0
Pillow>=10.0.0
scikit-image>=0.21.0

# GUI Framework
PyQt5>=5.15.0
pyqtgraph>=0.13.0

# Astronomy and Space Data
astropy>=5.3.0
astroquery>=0.4.6
fits2image>=0.1.0

# Image Processing and Analysis
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# API and Web
requests>=2.31.0
aiohttp>=3.8.0
fastapi>=0.100.0
uvicorn>=0.23.0

# Development and Testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.7.0
isort>=5.12.0
flake8>=6.0.0

# Documentation
sphinx>=7.1.0
sphinx-rtd-theme>=1.3.0

# Jupyter and Interactive Analysis
jupyter>=1.0.0
ipywidgets>=8.1.0
notebook>=7.0.0

# Utilities
tqdm>=4.65.0
click>=8.1.0
python-dotenv>=1.0.0
pyyaml>=6.0.0

# GPU acceleration (optional)
# cupy-cuda11x>=12.0.0  # Uncomment if using CUDA 11.x
# cupy-cuda12x>=12.0.0  # Uncomment if using CUDA 12.x
